pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.121.0",
    "multipart>=1.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
    "ulid-py>=1.1.0",
]

[dependency-groups]
dev = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "ruff>=0.14.4",
    "sqlalchemy>=2.0.44",
]
